+++ .github/CHANGELOG.md
@@ -3,12 +3,16 @@ This file contains update logs for this project. The top may contain a `Unreleas
 
 ## Unreleased
 ### Changed
-* Updated Velocity version to be compatible with Velocity 1.1.0 (use Maintenance 3.0.5 if you use an older version of Velocity)
+* Updated the Velocity version to be compatible with Velocity 1.1.0 (use Maintenance 3.0.5 if you use an older version of Velocity)
+* Added `maintenance schedule <minutes> <duration> command`
+  * This starts a timer for maintenance to be enabled, after which another timer will be started to disable it again (in contrast to `starttimer` and `endtimer`, which each only do one of)
+  * On proxies, you can use `maintenance schedule [server] <minutes> <duration> command` to schedule maintenance on a proxied server
 * Added command alias `mt` for the Maintenance command
 * Added secret argument to the whitelist add command: You can use `maintenance add <uuid> <name>` to add offline entries per command
 * The maintenance's base command is now properly registered with a permission on each platform, so that it is excluded from tabcomplection to unauthorized players
 * Some improvements to the French language file (thanks to @Aurelien30000)
 * API addition: Added MaintenanceReloadedEvent
+  * If you are using the MaintenancePlaceholderAddon, you need to [update it to 1.0.4](https://github.com/KennyTV/MaintenanceAddon/releases)
 ### Fixed
 * Fixed not setting custom singleserver kickmessages throwing warnings
 * Bungee: Fixed linebreaks in the playercount hovermessage for 1.16+ clients